fix(web): extract isBrowserNavigation util, fix file next param, migrate sso screen
- Move isBrowserNavigation from resolve-org-from-path.ts into
  apps/mesh/src/api/utils/browser-navigation.ts so routes can import
  a shared HTTP util without coupling to the org middleware
- Fix files.ts login redirect: use c.req.path (full /api/:org/files/:key)
  instead of /:slug so post-login navigation returns the user to the
  actual file rather than the org root
- Migrate SsoRequiredScreen to AccessScreenLayout, completing the branded
  gradient treatment across all five org access-gate screens
